# app/backend/cmd/api/docs/swagger.yaml
definitions:
  common.ErrorResponse:
    properties:
      error:
        description: 'example: user not found'
        type: string
    type: object
  handlers.BuildingResponse:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      stage:
        type: string
    type: object
...
  handlers.CreateBuildingRequest:
    properties:
      address:
        description: 'example: Невский пр., 1'
        type: string
      name:
        description: 'example: Дом на Невском'
        type: string
      stage:
        description: 'example: построено'
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      login:
        description: 'example: ivan123'
        type: string
      password:
        description: 'example: passw0rd'
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      lastname:
        description: 'example: Ivanov'
        type: string
      login:
        description: 'example: ivan123'
        type: string
      name:
        description: 'example: Ivan'
        type: string
      password:
        description: 'example: passw0rd'
        type: string
    type: object
...
  handlers.TokenResponse:
    properties:
      access_token:
        description: access token signed with server secret
        type: string
      expires_in:
        description: seconds until token expiration
        type: integer
      token_type:
        description: token type, usually "Bearer"
        type: string
    type: object
  handlers.UpdateBuildingRequest:
    properties:
      address:
        description: 'example: Новый адрес'
        type: string
      name:
        description: 'example: Новый дом'
        type: string
      stage:
        description: 'example: в_строительстве'
        type: string
    type: object
...
  handlers.UserResponse:
    properties:
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: this is documentation for buildefect API
  title: buildefect api
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Validate credentials and return access token with expiry seconds
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: invalid credentials
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Login and obtain JWT
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account (no JWT returned). Default role = "engineer".
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: invalid request body or missing fields
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: user already exists
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Register a user
      tags:
      - auth
  /api/buildings:
    get:
      consumes:
      - application/json
      description: Retrieve list of all buildings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BuildingResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List buildings
      tags:
      - buildings
    post:
      consumes:
      - application/json
      description: Create a new building record
      parameters:
      - description: Building payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BuildingResponse'
        "400":
          description: invalid request body or missing fields
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create building
      tags:
      - buildings
  /api/buildings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete building by id
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Successfully deleted building with id {id}
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete building
      tags:
      - buildings
    get:
      consumes:
      - application/json
      description: Retrieve building by numeric id
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BuildingResponse'
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: building not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get building by id
      tags:
      - buildings
    patch:
      consumes:
      - application/json
      description: Partially update building (name/address/stage)
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateBuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BuildingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update building
      tags:
      - buildings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"